+++ packages/programs/data/shared-log/test/role.spec.ts
@@ -1,11 +1,11 @@
-import { delay, waitFor, waitForResolved } from "@peerbit/time";
-import { EventStore } from "./utils/stores/event-store.js";
-import { TestSession } from "@peerbit/test-utils";
-import { Observer, ReplicationSegment, Replicator } from "../src/role.js";
 import { deserialize } from "@dao-xyz/borsh";
 import { Ed25519Keypair } from "@peerbit/crypto";
+import { SearchRequest } from "@peerbit/indexer-interface";
+import { TestSession } from "@peerbit/test-utils";
+import { delay, waitFor, waitForResolved } from "@peerbit/time";
+import { expect } from "chai";
 import { isMatured } from "../src/ranges.js";
-import { expect } from 'chai'
+import { EventStore } from "./utils/stores/event-store.js";
 
 describe(`role`, () => {
 	let session: TestSession;
@@ -21,11 +21,11 @@ describe(`role`, () => {
 							93, 202, 183, 249, 50, 240, 175, 84, 87, 239, 94, 92, 9, 207, 165,
 							88, 38, 234, 216, 0, 183, 243, 219, 11, 211, 12, 61, 235, 154, 68,
 							205, 124, 143, 217, 234, 222, 254, 15, 18, 64, 197, 13, 62, 84,
-							62, 133, 97, 57, 150, 187, 247, 215
+							62, 133, 97, 57, 150, 187, 247, 215,
 						]),
-						Ed25519Keypair
-					).toPeerId()
-				}
+						Ed25519Keypair,
+					).toPeerId(),
+				},
 			},
 			{
 				libp2p: {
@@ -35,11 +35,11 @@ describe(`role`, () => {
 							46, 27, 15, 0, 173, 134, 194, 249, 74, 80, 151, 42, 219, 238, 163,
 							44, 6, 244, 93, 0, 136, 33, 37, 186, 9, 233, 46, 16, 89, 240, 71,
 							145, 18, 244, 158, 62, 37, 199, 0, 28, 223, 185, 206, 109, 168,
-							112, 65, 202, 154, 27, 63, 15
+							112, 65, 202, 154, 27, 63, 15,
 						]),
-						Ed25519Keypair
-					).toPeerId()
-				}
+						Ed25519Keypair,
+					).toPeerId(),
+				},
 			},
 			{
 				libp2p: {
@@ -49,16 +49,16 @@ describe(`role`, () => {
 							12, 215, 160, 74, 43, 159, 235, 35, 84, 2, 7, 71, 15, 5, 210, 231,
 							155, 75, 37, 0, 15, 85, 72, 252, 153, 251, 89, 18, 236, 54, 84,
 							137, 152, 227, 77, 127, 108, 252, 59, 138, 246, 221, 120, 187,
-							239, 56, 174, 184, 34, 141, 45, 242
+							239, 56, 174, 184, 34, 141, 45, 242,
 						]),
-						Ed25519Keypair
-					).toPeerId()
-				}
-			}
+						Ed25519Keypair,
+					).toPeerId(),
+				},
+			},
 		]);
 		await session.connect([
 			[session.peers[0], session.peers[1]],
-			[session.peers[1], session.peers[2]]
+			[session.peers[1], session.peers[2]],
 		]);
 		await session.peers[1].services.blocks.waitFor(session.peers[0].peerId);
 		await session.peers[2].services.blocks.waitFor(session.peers[1].peerId);
@@ -68,22 +68,26 @@ describe(`role`, () => {
 		await session.stop();
 	});
 
-	beforeEach(async () => {
-		db1 = await session.peers[0].open(new EventStore<string>());
-	});
+	beforeEach(async () => {});
 
 	afterEach(async () => {
-		await db1?.drop();
-		await db2?.drop();
+		if (db1?.closed === false) {
+			await db1?.drop();
+		}
+		if (db2?.closed === false) {
+			await db2?.drop();
+		}
 	});
 
 	it("none", async () => {
+		db1 = await session.peers[0].open(new EventStore<string>());
+
 		db2 = (await EventStore.open<EventStore<string>>(
 			db1.address!,
 			session.peers[1],
 			{
-				args: { role: "observer" }
-			}
+				args: { replicate: false },
+			},
 		))!;
 
 		await db1.waitFor(session.peers[1].peerId);
@@ -91,43 +95,44 @@ describe(`role`, () => {
 		await db1.add("hello");
 		await db2.add("world");
 
-		await waitFor(() => db1.log.log.values.length === 2); // db2 can write ...
+		await waitFor(() => db1.log.log.length === 2); // db2 can write ...
 		expect(
-			(await db1.log.log.values.toArray()).map(
-				(x) => x.payload.getValue().value
-			)
+			(await db1.log.log.toArray()).map((x) => x.payload.getValue().value),
 		).to.have.members(["hello", "world"]);
-		expect(db2.log.log.values.length).equal(1); // ... but will not receive entries
+		expect(db2.log.log.length).equal(1); // ... but will not receive entries
 	});
 
 	describe("observer", () => {
 		it("can update", async () => {
+			db1 = await session.peers[0].open(new EventStore<string>());
+
 			expect(
-				(db1.log.node.services.pubsub as any)["subscriptions"].get(db1.log.rpc.topic)
-					.counter
+				(db1.log.node.services.pubsub as any)["subscriptions"].get(
+					db1.log.rpc.topic,
+				).counter,
 			).equal(1);
+			expect([...(await db1.log.getReplicators())]).to.deep.equal([
+				db1.node.identity.publicKey.hashcode(),
+			]);
+			expect(await db1.log.isReplicating()).to.be.true;
+			await db1.log.replicate(false);
+			expect(await db1.log.isReplicating()).to.be.false;
 			expect(
-				db1.log
-					.getReplicatorsSorted()
-					?.toArray()
-					?.map((x) => x.publicKey.hashcode())
-			).to.deep.equal([db1.node.identity.publicKey.hashcode()]);
-			expect(db1.log.role).to.be.instanceOf(Replicator);
-			await db1.log.updateRole("observer");
-			expect(db1.log.role).to.be.instanceOf(Observer);
-			expect(
-				(db1.log.node.services.pubsub as any)["subscriptions"].get(db1.log.rpc.topic)
-					.counter
+				(db1.log.node.services.pubsub as any)["subscriptions"].get(
+					db1.log.rpc.topic,
+				).counter,
 			).equal(1);
 		});
 
 		it("observer", async () => {
+			db1 = await session.peers[0].open(new EventStore<string>());
+
 			db2 = (await EventStore.open<EventStore<string>>(
 				db1.address!,
 				session.peers[1],
 				{
-					args: { role: "observer" }
-				}
+					args: { replicate: false },
+				},
 			))!;
 
 			await db1.waitFor(session.peers[1].peerId);
@@ -135,24 +140,39 @@ describe(`role`, () => {
 			await db1.add("hello");
 			await db2.add("world");
 
-			await waitFor(() => db1.log.log.values.length === 2); // db2 can write ...
+			await waitFor(() => db1.log.log.length === 2); // db2 can write ...
 			expect(
-				(await db1.log.log.values.toArray()).map(
-					(x) => x.payload.getValue().value
-				)
+				(await db1.log.log.toArray()).map((x) => x.payload.getValue().value),
 			).to.have.members(["hello", "world"]);
-			expect(db2.log.log.values.length).equal(1); // ... but will not receive entries
+			expect(db2.log.log.length).equal(1); // ... but will not receive entries
 		});
 	});
 
 	describe("replictor", () => {
+		it("fixed", async () => {
+			db1 = await session.peers[0].open(new EventStore<string>(), {
+				args: {
+					replicate: {
+						offset: 0.7,
+						factor: 0.5,
+					},
+				},
+			});
+			let ranges = await db1.log.getMyReplicationSegments();
+			expect(ranges).to.have.length(1);
+			expect(ranges[0].toReplicationRange().offset).to.closeTo(0.7, 0.000001);
+			expect(ranges[0].toReplicationRange().factor).to.closeTo(0.5, 0.000001);
+		});
+
 		it("dynamic by default", async () => {
+			db1 = await session.peers[0].open(new EventStore<string>());
+
 			db2 = (await EventStore.open<EventStore<string>>(
 				db1.address!,
-				session.peers[1]
+				session.peers[1],
 			))!;
 			const roles: any[] = [];
-			db2.log.events.addEventListener("role", (change) => {
+			db2.log.events.addEventListener("replication:change", (change) => {
 				if (
 					change.detail.publicKey.equals(session.peers[1].identity.publicKey)
 				) {
@@ -164,18 +184,20 @@ describe(`role`, () => {
 		});
 
 		it("passing by string evens by default", async () => {
+			db1 = await session.peers[0].open(new EventStore<string>());
+
 			db2 = await EventStore.open<EventStore<string>>(
 				db1.address!,
 				session.peers[1],
 				{
 					args: {
-						role: "replicator"
-					}
-				}
+						replicate: true,
+					},
+				},
 			);
 
 			const roles: any[] = [];
-			db2.log.events.addEventListener("role", (change) => {
+			db2.log.events.addEventListener("replication:change", (change) => {
 				if (
 					change.detail.publicKey.equals(session.peers[1].identity.publicKey)
 				) {
@@ -191,94 +213,116 @@ describe(`role`, () => {
 
 			const db1 = await session.peers[0].open(store.clone(), {
 				args: {
-					role: {
-						type: "replicator",
-						factor: 1
-					}
-				}
+					replicate: {
+						factor: 1,
+					},
+				},
 			});
 			const db2 = await session.peers[1].open(store.clone(), {
 				args: {
-					role: {
-						type: "replicator",
-						factor: 1
-					}
-				}
+					replicate: {
+						factor: 1,
+					},
+				},
 			});
-			db2.log.getDefaultMinRoleAge = () => 3e3;
+			db2.log.getDefaultMinRoleAge = () => Promise.resolve(3e3);
 			const t0 = +new Date();
 			await db2.log.waitForReplicator(db1.node.identity.publicKey);
 			const t1 = +new Date();
-			expect(t1 - t0).greaterThan(db2.log.getDefaultMinRoleAge());
+			expect(t1 - t0).greaterThan(await db2.log.getDefaultMinRoleAge());
 		});
 		describe("getDefaultMinRoleAge", () => {
+			it("if not replicating, min role age is 0", async () => {
+				const store = new EventStore<string>();
+
+				await session.peers[0].open(store.clone(), {
+					args: {
+						replicate: {
+							factor: 1,
+						},
+					},
+				});
+				const db2 = await session.peers[1].open(store.clone(), {
+					args: {
+						replicate: false,
+					},
+				});
+				await waitForResolved(async () =>
+					expect((await db2.log.getReplicators()).size).to.equal(1),
+				);
+				expect(await db2.log.getDefaultMinRoleAge()).equal(0);
+			});
+
 			it("oldest is always mature", async () => {
 				const store = new EventStore<string>();
 
 				const db1 = await session.peers[0].open(store.clone(), {
 					args: {
-						role: {
-							type: "replicator",
-							factor: 1
-						}
-					}
+						replicate: {
+							factor: 1,
+						},
+					},
 				});
-				const tsm = 100;
+				const tsm = 1000;
 
 				await delay(tsm);
+
 				const db2 = await session.peers[1].open(store.clone(), {
 					args: {
-						role: {
-							type: "replicator",
-							factor: 1
-						}
-					}
+						replicate: {
+							factor: 1,
+						},
+					},
 				});
-				await waitForResolved(() =>
-					expect(db1.log.getReplicatorsSorted()?.length).equal(2)
+				await waitForResolved(async () =>
+					expect(await db1.log.replicationIndex?.getSize()).equal(2),
 				);
-				await waitForResolved(() =>
-					expect(db2.log.getReplicatorsSorted()?.length).equal(2)
+				await waitForResolved(async () =>
+					expect(await db2.log.replicationIndex?.getSize()).equal(2),
 				);
 
-				const db1MinRoleAge = db1.log.getDefaultMinRoleAge();
-				const db2MinRoleAge = db2.log.getDefaultMinRoleAge();
+				const db1MinRoleAge = await db1.log.getDefaultMinRoleAge();
+				const db2MinRoleAge = await db2.log.getDefaultMinRoleAge();
 
 				expect(db1MinRoleAge - db2MinRoleAge).lessThanOrEqual(1); // db1 sets the minRole age because it is the oldest. So both dbs get same minRole age limit (including some error margin)
-				expect(db1MinRoleAge).greaterThan(tsm);
+
 				const now = +new Date();
 
 				// Mature because if "first"
 				let selfMatured = isMatured(
-					db1.log.role as Replicator,
+					(await db1.log.getMyReplicationSegments())[0],
 					now,
-					db1.log.getDefaultMinRoleAge()
+					await db1.log.getDefaultMinRoleAge(),
 				);
 				expect(selfMatured).to.be.true;
-				await waitForResolved(() =>
+
+				await waitForResolved(async () => {
+					const minRoleAge = await db1.log.getDefaultMinRoleAge();
 					expect(
-						db1.log
-							.getReplicatorsSorted()
-							?.toArray()
-							.filter((x) =>
-								isMatured(x.role, now, db1.log.getDefaultMinRoleAge())
-							)
-							.map((x) => x.publicKey.hashcode())
-					).to.deep.equal([db1.node.identity.publicKey.hashcode()])
-				);
+						(await db1.log.replicationIndex.query(new SearchRequest())).results
+							.map((x) => x.value)
+							.filter((x) => isMatured(x, now, minRoleAge))
+							.map((x) => x.hash),
+					).to.deep.equal([db1.node.identity.publicKey.hashcode()]);
+				});
 
 				// assume other nodes except me are mature if the open before me
 				selfMatured = isMatured(
-					db2.log.role as Replicator,
+					(await db2.log.getMyReplicationSegments())[0],
 					now,
-					db2.log.getDefaultMinRoleAge()
+					await db2.log.getDefaultMinRoleAge(),
 				);
 				expect(selfMatured).to.be.false;
+
+				const minRoleAge = await db2.log.getDefaultMinRoleAge();
 				expect(
-					db2.log
-						.getReplicatorsSorted()
-						?.toArray()
-						.map((x) => isMatured(x.role, now, db2.log.getDefaultMinRoleAge()))
+					(
+						await db2.log.replicationIndex.query(
+							new SearchRequest({ fetch: 0xffffffff }),
+						)
+					).results
+						.map((x) => x.value)
+						.map((x) => isMatured(x, now, minRoleAge)),
 				).to.have.members([false, true]);
 			});
 
@@ -286,7 +330,7 @@ describe(`role`, () => {
 		});
 	});
 });
-
+/* 
 describe("segment", () => {
 	describe("overlap", () => {
 		it("non-wrapping", () => {
@@ -315,7 +359,7 @@ describe("segment", () => {
 			expect(s2.overlaps(s1)).to.be.true;
 		});
 	});
-});
+}); */
 
 /* it("encrypted clock sync write 1 entry replicate false", async () => {
 	await waitForPeers(session.peers[1], [client1.id], db1.address.toString());
@@ -344,9 +388,9 @@ describe("segment", () => {
 	// Now the db2 will request sync clocks even though it does not replicate any content
 	await db2.add("world");
 
-	await waitFor(() => db1.store.oplog.values.length === 2);
+	await waitFor(() => db1.store.oplog.length === 2);
 	expect(
-		db1.store.oplog.values.toArray().map((x) => x.payload.getValue().value)
+		db1.store.oplog.toArray().map((x) => x.payload.getValue().value)
 	).to.have.members(["hello", "world"]);
-	expect(db2.store.oplog.values.length).equal(1);
+	expect(db2.store.oplog.length).equal(1);
 }); */
